+++ src/main/java/net/messer/mystical_index/item/custom/page/attribute/LinksPage.java
@@ -5,7 +5,7 @@ import net.minecraft.item.ItemStack;
 
 public class LinksPage extends IndexingTypePage.IndexingAttributePage {
     @Override
-    public int getLinksMultiplier(ItemStack page) {
+    public double getLinksMultiplier(ItemStack page) {
         return 2;
     }
